requirements.txt
djongo
Django
requests
pymongo==3.12.1
sqlparse==0.2.4
pytz
djangorestframework